requirements.txt
polars>=1.10.0
xlsxwriter>=3.2.0
openpyxl>=3.1.2
fastexcel>=0.11.5
pandas>=2.2.0
google-generativeai>=0.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
